backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.2
pillow==10.1.0
scipy==1.11.4
python-dotenv==1.0.0